pyproject.toml
[tool.poetry]
name = "cdsa-langchain-openai-chatbot"
version = "0.1.0"
description = ""
authors = ["Surya_Pratap"]
license = "None"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.13"
chainlit = "^1.0.401"
langchain-openai = "^0.1.1"
streamlit = "^1.32.2"
langchain-experimental = "^0.0.55"
langchain-core = "^0.1.36"
python-dotenv = "^1.0.1"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
